% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clanEGSnrc}
\alias{clanEGSnrc}
\alias{clanOptim}
\title{Version function for the clanOptim library}
\usage{
clanEGSnrc()
}
\value{
A list of information about the version and functions within clanEGSnrc.
}
\description{
Version function for the clanOptim library.
}
\author{
Claus E. Andersen
}

