% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.flurznrc.lst}
\alias{read.flurznrc.lst}
\title{Read fluence spactre in lst file produced by the EGSnrc usercode flurznrc}
\usage{
df1  <- read.egsnrc( file.path="C:/data/projects/Theoretical-Dosimetry",
file.name="mcclan-10410-thin-slab.spectra",
file.name.save="mcclan-10410-thin-slab",
read.vec=c(1,3),
trace=!TRUE)
}
\arguments{
\item{file.path:}{Path to folder where file to be read is located.}

\item{file.name:}{Name of file to be read}

\item{file.name.save:}{If "" then nothing is saved. Otherwise files
are saved for each individual plot. Not implemented.}

\item{read.vec:}{Vector of element that should be processes. c(1,5) means
that only plot 1 and 5 should be processed. 1:99 means that
all plots up to 99 should be processed.}

\item{plot.wanted:}{Not implemented.}

\item{trace:}{TRUE means that the function becomes verbose. FALSE means silience.}
}
\value{
A single dataframe (which can be stacked, if more data files of the same type are read).
}
\description{
Read fluence spactre in lst file produced by the EGSnrc usercode flurznrc.
See body of function for further details.
}
\author{
Claus E. Andersen
}

